+++ HISTORY.md
@@ -3,6 +3,9 @@
 ## x.x.x (not yet released)
 * Allow a Hint to be the first argument to Hint::build().
 * Make `Hint::*` constants into a `HintType` enumeration.
+* Remove old hint back-compatibility that allowed suffixing `[]` to a
+  class-name string.  Use `Hint::build(Foo::class, HintType::LIST)`
+  instead.
 
 ## 4.0.0 (2025-08-04)
 * Add `Hint::ONLY_FOR_DECODE` to support forward-compatibility of